example: Remove Java dependency
The mraa recipe from iot-middleware comes with a mraa-java and upm-java
packages, but we do not need them and we do not install them. However,
the packages create build-time dependency on Java, which is prolonging
the build at best. At worst, it is importing build breakages due to
meta-java remaining fragile.

Remove this dependency by removing the mraa-java package

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>